% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.R
\name{ATE}
\alias{ATE}
\title{Estimating the average treatment effect given non-random attrition}
\usage{
ATE(regressionFormula, instrumentFormula, data, p_W_Formula = R ~ .,
  p_W_Method = binomial(link = logit), PiFormula = D ~ .,
  PiMethod = binomial(link = logit), nBoots = 1000, quantiles = c(0.05,
  0.95), effectType = "All", nCores = 1)
}
\arguments{
\item{regressionFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{regressionFormula}
is the model which will be used to estimate the average treatment effect accounting for 
non-random attrition. Formula must be of the form Y ~ D + ..., where Y is the outcome, 
D is the treatment, and ... represents any additional covariates.  The first covariate
in the formula MUST be the treatment, D.}

\item{instrumentFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{instrumentFormula}
contains the instruments to be used to calculate the weights which account for non-random
atrrition.  Formula must be of the form ~ Z1 + ..., where Z1 represents
some instrumental variable, and ... represents any additional instruments.}

\item{data}{A data frame which contains all variables to be used in both 
\code{regressionFormula} and \code{instrumentFormula}.}

\item{p_W_Formula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{p_W_Formula}
is the model which estimates subjects' response propensity, which is used in generating
the weights which account for non-random attrition.  By default, the formula includes
all covariates included in \code{regressionFormula}.}

\item{p_W_Method}{A character string which indicates the the error distribution and 
link function to be used in the generalized additive model which estimates subjects'
response propensity.  By default, \code{p_W_Formula} is set to 
\code{binomial(link = logit)}, which fits a logistic regression model.  Any family
function that \code{gam} accepts can be specified here.}

\item{PiFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{p_W_Formula}
is the model which estimates subjects' treatment propensity, which is used in generating
the weights which account for non-random attrition.  By default, the formula includes
all covariates included in \code{regressionFormula}.}

\item{PiMethod}{A character string which indicates the the error distribution and 
link function to be used in the generalized additive model which estimates subjects'
treatment propensity.  By default, \code{p_W_Formula} is set to 
\code{binomial(link = logit)}, which fits a logistic regression model.  Any family
function that \code{gam} accepts can be specified here.}

\item{nBoots}{Numeric value defining the number of bootstrap samples; the default number
of bootstrap replications is 1000.}

\item{quantiles}{Vector of two numeric values between 0 and 1 specifying the quantiles
at which estimates of the average treatment effect will be returned.  
By default, \code{quantiles} are set to 0.05 and 0.95, which corresponds with a 90\%
confidence interval.}

\item{effectType}{Character string indicating the type of effect types to be calculated
and returned.  Valid inputs are 'Respondent,' which calculates the average treatment
effect only among respondents; 'Population,' which calculates the average treatment effect
among all subjects; and 'All,' which calculates both the average treatment effect only
only among respondents, and among all subjects.}

\item{nCores}{Numeric value indicating the number of cores to allocate to running the
function.  See \code{parallel} for details.}
}
\value{
A list of five elements containing the following:
 \item{MeanEst}{A list containing the mean estimated coefficients for the regression
 used to estimate the average treatment effect.  List will contain the estimates
 specified by the option \code{effectType}.}
 \item{MedianEst}{A list containing the median estimated coefficients for the regression
 used to estimate the average treatment effect.  List will contain the estimates
 specified by the option \code{effectType}.}
 \item{SE}{A list containing the mean estimated standard errors of the estimates
 for the regression used to estimate the average treatment effect.  List will contain 
 the standard errors for the estimates specified by the option \code{effectType}.}
 \item{Quantiles}{A list containing the coefficient estimates for the regression
 used to estimate the average treatment effect that correspond with the quantiles
 specified by \code{quantile}.  List will contain the quantiles specified by the option
  \code{effectType}.}
 \item{Matrix}{A matrix containing the coefficient estimates from each bootstrapped
 sample.  Each row of the matrix corresponds with one of the coefficients estimated
 in the model, and the coefficient estimates in each column are the estimates obtained
 in the bootstrapped sample corresponding with the column index.}
}
\description{
\code{ATE} estimates the average treatment effect under conditions of
 non-random attrition.
}
\details{
The function estimates the average treatment effect (ATE) under non-random attrition,
and uses bootstrapping to estimate mean value, median, and standard errors 
based on the bootstrapped replications.
}
\references{
Huber, Martin (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh
}
