% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapDelta.R, R/ATE.R
\name{bootstrapDelta}
\alias{bootstrapDelta}
\alias{ATE}
\title{Estimating standard errors for ATEs given non-random attrition}
\usage{
bootstrapDelta(regressionFormula, instrumentFormula, data, p_W_Formula = R ~
  ., p_W_Method = binomial(link = logit), PiFormula = D ~ .,
  PiMethod = binomial(link = logit), nBoots = 1000, quantiles = c(0.05,
  0.95), effectType = "Population", nCores = 1)

ATE(regressionFormula, instrumentFormula, data, p_W_Formula = R ~ .,
  p_W_Method = binomial(link = logit), PiFormula = D ~ .,
  PiMethod = binomial(link = logit), nBoots = 1000, quantiles = c(0.05,
  0.95), effectType = "All", nCores = 1)
}
\arguments{
\item{regressionFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{regressionFormula}
is the model which will be used to estimate the ATE after accounting for non-random
attrition. Formula must be of the form Y ~ D + ..., where Y is the outcome, D is the treatment,
and ... represents any set of other covariates.}

\item{instrumentFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{instrumentFormula}
is the model which estimates the weights used in the model which estimates the ATE.
Formula must be of the form ~ Z1 + ..., where Z1 represents some instrumental variable,
and ... represents any other desired instrumental variables.}

\item{data}{A data frame which contains all variables to be used in both 
\code{regressionFormula} and \code{instrumentFormula}.}

\item{p_W_Formula}{The regression formula (of class \code{formula}) for calculating the response propensity probabilities.
By default this formula is of the form R ~ ., where R represents the presence of a response 
and . represents treatment, desired covariates, and the instrument. Users are strongly discouraged 
from manipulating this formula.}

\item{p_W_Method}{The regression method for calculating the response propensity probabilities.
By default this is binomial(link = logit). Other valid methods include those that function in the 
family argument to the \code{gam} function.}

\item{PiFormula}{The regression formula (of class \code{formula}) for calculating the treatment propensity probabilities.
By default this formula is of the form D ~ ., where D represents treatment and . represents
desired covariates and the response propensity probabilities. Users are strongly discouraged from
manipulating this formula.}

\item{PiMethod}{The regression method for calculating the treatment propensity probabilities.
By default this is binomial(link = logit). Other valid methods include those that function in the 
 family argument to the \code{gam} or function.}

\item{nBoots}{Numeric value defining the number of bootstrap samples; the default number
of bootstrap replications is 1,000.}

\item{quantiles}{A vector of percentiles for which the bootstrapped quantiles will be returned.
By default the function returns the quantiles corresponding to the 5th and 95th percentiles.}

\item{effectType}{A string of either 'Population', 'Respondent', or 'Both', corresponding
to the desired ATE calculation. By default, the function calculates the ATE on the population.}

\item{nCores}{}

\item{regressionFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{regressionFormula}
is the model which will be used to estimate the average treatment effect accounting for 
non-random attrition. Formula must be of the form Y ~ D + ..., where Y is the outcome, 
D is the treatment, and ... represents any additional covariates.}

\item{instrumentFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{instrumentFormula}
is the model which estimates the weights used in the model which estimates the average
treatment effect.  Formula must be of the form ~ Z1 + ..., where Z1 represents
some instrumental variable, and ... represents any additional covariates.}

\item{data}{A data frame which contains all variables to be used in both 
\code{regressionFormula} and \code{instrumentFormula}.}

\item{p_W_Formula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{p_W_Formula}
is the model which estimates subjects' response propensity, which is used in generating
the weights which account for non-random attrition.  By default, the formula includes
all covariates in the specified data frame.}

\item{p_W_Method}{A character string which indicates the the error distribution and 
link function to be used in the generalized additive model which estimates subjects'
response propensity.  By default, \code{p_W_Formula} is set to 
\code{binomial(link = logit)}, which fits a logistic regression model.  Any family
function that \code{gam} accepts can be specified here.}

\item{PiFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{p_W_Formula}
is the model which estimates subjects' treatment propensity, which is used in generating
the weights which account for non-random attrition.  By default, the formula includes
all covariates in the specified data frame.}

\item{PiMethod}{A character string which indicates the the error distribution and 
link function to be used in the generalized additive model which estimates subjects'
treatment propensity.  By default, \code{p_W_Formula} is set to 
\code{binomial(link = logit)}, which fits a logistic regression model.  Any family
function that \code{gam} accepts can be specified here.}

\item{nBoots}{Numeric value defining the number of bootstrap samples; the default number
of bootstrap replications is 1000.}

\item{quantiles}{Vector of two numeric values between 0 and 1 specifying the quantiles
at which estimates of the average treatment effect will be returned.  
By default, \code{quantiles} are set to 0.05 and 0.95, which corresponds with a 90\%
confidence interval.}

\item{effectType}{Character string indicating the type of effect types to be calculated
and returned.  Valid inputs are 'Respondent,' which calculates the average treatment
effect only among respondents; 'Population,' which calculates the average treatment effect
among all subjects; and 'All,' which calculates both the average treatment effect only
only among respondents, and among all subjects.}

\item{nCores}{Numeric value indicating the number of cores to allocate to running the
function.  See \code{parallel} for details.}
}
\value{
A list of six lists. If effectType = 'Both', each list contains two lists, otherwise
each element consists of only one list.
 \item{MeanEst}{A list containing a numeric vector of mean estimates for the effects of treatment and
 covariates of interest.}
 \item{MedianEst}{A list containing a numeric vector of median estimates for the effects of treatment and
 covariates of interest.}
 \item{SE}{A list containing a numeric vector of standard deviations for the estimated effects of treatment and
 covariates of interest.}
 \item{Quantiles}{A list containing a matrix of quantiles (rows) corresponding to the estimates
 for the treatment and covariate effects (columns). By defualt it returns quantiles corresponding to
 the 5th and 95th percentiles.}
 \item{Matrix}{A list containing a matrix of bootstrapped coefficient estimates, where each
 row corresponds to a covariate, and each column corresponds to a bootstrap iteration.}
 \item{Data}{A list containing a list of bootstrapped data frames.}

A list of five elements containing the following:
 \item{MeanEst}{A list containing the mean estimated coefficients for the regression
 used to estimate the average treatment effect.  List will contain the estimates
 specified by the option \code{effectType}.}
 \item{MedianEst}{A list containing the median estimated coefficients for the regression
 used to estimate the average treatment effect.  List will contain the estimates
 specified by the option \code{effectType}.}
 \item{SE}{A list containing the mean estimated standard errors of the estimates
 for the regression used to estimate the average treatment effect.  List will contain 
 the standard errors for the estimates specified by the option \code{effectType}.}
 \item{Quantiles}{A list containing the coefficient estimates for the regression
 used to estimate the average treatment effect that correspond with the quantiles
 specified by \code{quantile}.  List will contain the quantiles specified by the option
  \code{effectType}.}
 \item{Matrix}{A matrix containing the coefficient estimates from each bootstrapped
 sample.  Each row of the matrix corresponds with one of the coefficients estimated
 in the model, and the coefficient estimates in each column are the estimates obtained
 in the bootstrapped sample corresponding with the column index.}
}
\description{
\code{bootstrapDelta} estimates standard errors of the average treatment effect (ATE)
under conditions of non-random attrition.

\code{ATE} estimates the average treatment effect under conditions of
 non-random attrition.
}
\details{
The function estimates standard errors for the average treatment effect (ATE)
after accounting for non-random attrition. After drawing random samples with 
replacement from the provided data (by default, 1000 samples), the function 
calculates the sample mean, median, and standard errors 
based on the bootstrapped estimates of the ATE.

The function estimates the average treatment effect (ATE) under non-random attrition,
and uses bootstrapping to estimate mean value, median, and standard errors 
based on the bootstrapped replications.
}
\references{
Huber, Martin (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.

Huber, Martin (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh

Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh
}
