% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeights.R
\name{calculateWeights}
\alias{calculateWeights}
\title{Calculating weights for treatment and control group under non-random attrition}
\usage{
calculateWeights(modelData, instrumentData, p_W_Formula = R ~ .,
  p_W_Method = binomial(link = logit), PiFormula = D ~ .,
  PiMethod = binomial(link = logit))
}
\arguments{
\item{modelData}{A matrix with columns of \code{Y}, \code{D}, \code{X}, where 
\code{Y} indicates outcome variable which may or may not include NA (non-response), 
\code{D} indicates treatment indicator, either 1 (treatment) or 0 (non-treatment),
\code{X} indicates observed covariates, and \code{Z} indicates instrumental variable. 
\code{modelData} must structured with \code{Y}, \code{D}, \code{X}. 
Otherwise, the incorrect output will be produced.}

\item{instrumentData}{A matrix with a column of \code{Z}, indicating instrumental variable. 
\code{Z} is a predictor of response variable (R) of 1 if \code{Y} is observed and 0 otherwise (attribution).
It needs to contain at least one continuous element and not to have a direct effect on the outcome (\code{Y}).}
}
\value{
A dataframe of estimated weights for treated and untreated(control) group.
 \item{column1}{A dataframe of estimated weights for treated and untreated group, 
 returned by products of predicted values which are calculated by \code{method}.}
}
\description{
\code{calculateWeights} estimates the weights for treatment and control group 
in order to obtain internally and externally valid estimates for the average treatment effect (ATE) 
in cases of attrition.
}
\details{
The function estimates the response propensity score and the treatment propensity score 
to reweigh the observations and correct for attrition. 
The response propensity score is defined as $p(W) = Pr(R=1 | \code{X},\code{D}, \code{Z})$ 
where $R$ denotes to the binary response variable, being 1 if \code{Y} is observed and 0 otherwise (attrition), 
and the treatment propensity score, defined as $\pi(\code{X}, p(W)) = Pr(\code{D}=1 | \code{X}, p(W), R=1)$.
Response and treatment propensity scores are multiplied for treatment and control group
in order to adjust for differences in the distributions of X and p(W) between treated and nontreated. 
For more information on the method, see Huber (2012) below.
}
\references{
Huber (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh
}
