% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedAttrition}
\alias{SimulatedAttrition}
\title{Data with simulated attrition on unobservables.}
\format{A data frame with 1000 rows and 4 variables:
\describe{
  \item{Y}{Response variable generated from the equation 
  Y = 1*Treatment + 1*Covariate + 0.25*Treatment*Covariate + U;
  where U is an unobserved error term. 
  Missingness in Y is generated by the equation R = -1*Treatment + 1*Covariate + 1*Instrument + V;
  where V is an unobserved error term that is correlated with U.
  Specifically, U and V are drawn randomly from a multivariate standard normal distribution with a covariance of 0.8.}
  \item{D}{Treatment assignment drawn randomly from a binomial distribution with equal probability.}
  \item{X}{Covariate generated from random draws on a uniform distribution from -1 to 1.}
  \item{Z}{Instrument generated from random draws on a uniform distribution from -1 to 2.}
  ...
}}
\usage{
SimulatedAttrition
}
\description{
A dataset containing simulated data for computation with attritR.
}
\keyword{datasets}
