% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapDelta.R
\name{bootstrapDelta}
\alias{bootstrapDelta}
\title{Estimating standard errors for ATEs given non-random attrition}
\usage{
bootstrapDelta(regressionFormula, instrumentFormula, data, p_W_Formula = R ~
  ., p_W_Method = binomial(link = logit), PiFormula = D ~ .,
  PiMethod = binomial(link = logit), nBoots = 1000, quantiles = c(0.05,
  0.95), prop = "All", nCores = 1)
}
\arguments{
\item{regressionFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{regressionFormula}
is the model which will be used to estimate the ATE after accounting for non-random
attrition. Formula must be of the form Y ~ D + ..., where Y is the outcome, D is the treatment,
and ... represents any set of other covariates.}

\item{instrumentFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{instrumentFormula}
is the model which estimates the weights used in the model which estimates the ATE.
Formula must be of the form ~ Z1 + ..., where Z1 represents some instrumental variable,
and ... represents any other desired instrumental variables.}

\item{data}{A data frame which contains all variables to be used in both 
\code{regressionFormula} and \code{instrumentFormula}.}

\item{p_W_Formula}{The regression formula (of class \code{formula}) for calculating the response propensity probabilities.
By default this formula is of the form R ~ ., where R represents the presence of a response 
and . represents treatment, desired covariates, and the instrument. Users are strongly discouraged 
from manipulating this formula.}

\item{p_W_Method}{The regression method for calculating the response propensity probabilities.
By default this is binomial(link = logit). Other valid methods include those that function in the 
family argument to the \code{gam} function.}

\item{PiFormula}{The regression formula (of class \code{formula}) for calculating the treatment propensity probabilities.
By default this formula is of the form D ~ ., where D represents treatment and . represents
desired covariates and the response propensity probabilities. Users are strongly discouraged from
manipulating this formula.}

\item{PiMethod}{The regression method for calculating the treatment propensity probabilities.
By default this is binomial(link = logit). Other valid methods include those that function in the 
 family argument to the \code{gam} or function.}

\item{nBoots}{Numeric value defining the number of bootstrap samples; the default number
of bootstrap replications is 1,000.}

\item{quantiles}{A vector of percentiles for which the bootstrapped quantiles will be returned.
By default the function returns the quantiles corresponding to the 5th and 95th percentiles.}

\item{prop}{A string of either 'Population', 'Respondent', or 'All', corresponding
to the desired ATE calculation. By default, the function calculates the ATE on the population.}

\item{nCores}{}
}
\value{
A list of six lists. If prop = 'Both', each list contains two lists, otherwise
each element consists of only one list.
 \item{Means}{A list containing a numeric vector of mean estimates for the effects of treatment and
 covariates of interest.}
 \item{Medians}{A list containing a numeric vector of median estimates for the effects of treatment and
 covariates of interest.}
 \item{SD}{A list containing a numeric vector of standard deviations for the estimated effects of treatment and
 covariates of interest.}
 \item{Quants}{A list containing a matrix of quantiles (rows) corresponding to the estimates
 for the treatment and covariate effects (columns). By defualt it returns quantiles corresponding to
 the 5th and 95th percentiles.}
 \item{Matrix}{A list containing a matrix of bootstrapped coefficient estimates, where each
 row corresponds to a covariate, and each column corresponds to a bootstrap iteration.}
 \item{Data}{A list containing a list of bootstrapped data frames.}
}
\description{
\code{bootstrapDelta} estimates standard errors of the average treatment effect (ATE)
under conditions of non-random attrition.
}
\details{
The function estimates standard errors for the average treatment effect (ATE)
after accounting for non-random attrition. After drawing random samples with 
replacement from the provided data (by default, 1000 samples), the function 
calculates the sample mean, median, and standard errors 
based on the bootstrapped estimates of the ATE.
}
\references{
Huber, Martin (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh
}
