% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipwlm.R
\name{ipwlm}
\alias{ipwlm}
\title{Estimating treatment effects given non-random attrition}
\usage{
ipwlm(regression_formula, treatment, instrument = NULL, data,
  effect_type, attrition_type, response_weight_formula = response ~ .,
  response_weight_method = binomial(link = logit),
  treatment_weight_formula = data[treatment] ~ .,
  treatment_weight_method = binomial(link = logit),
  n_bootstraps = 1000, quantiles = c(0.05, 0.95), n_cores = 1)
}
\arguments{
\item{regression_formula}{An object of class \code{formula} (or one that can 
be coerced to that class); a symbolic description of the model to be fitted.  
\code{regression_formula} is the model which will be used to estimate the 
average treatment effect accounting for non-random attrition. Formula must be 
of the form Y ~ D + ..., where Y is the outcome, D is the treatment, and ... 
represents any additional covariates, including an instrumental variable when 
weighting for attrition on unobservables.}

\item{treatment}{A string specifying the treatment variable in the regression.
The string must match the name of the treatment variable exactly as it is 
passed into \code{regression_formula}.}

\item{instrument}{A string specifying the instrumental variable in the 
regression, if one is included. The string must match the name of the 
treatment variable exactly as it is passed into \code{regression_formula}. If 
no instrumental variable is used, the argument input should remain \code{NULL}.}

\item{data}{A data.frame which contains all variables specified in
\code{regression_formula}.}

\item{effect_type}{A string specifying the type of treatment effect to be 
estimated. Input must be one of \code{"respondent"} or \code{"population"}, 
which refer to the average treatment effect among respondents (ATE|R = 1) and 
the average treatment effect among the population (ATE), respectively.}

\item{attrition_type}{A string specifying the assumed effect of attrition. 
Input must be one of \code{"treatment"}, \code{"observable"}, or 
\code{"unobservable"}. These inputs refer, in order, to attrition due to 
treatment alone, attrition due to treatment and observable confounders, and 
attrition due to treatment, observable confounders, and unobservable 
confounders. If \code{"unobservable"} is specified, an instrumental variable
must be included in \code{regression_formula} and its name must be noted in 
\code{instrument}.}

\item{response_weight_formula}{An object of class \code{formula} (or one that 
can be coerced to that class); a symbolic description of the model to be 
fitted. \code{response_weight_formula} specifies the model for estimating 
respondents' response propensity, which is used in weighting treatment effect
estimates that account for non-random attrition. By default, the formula 
includes all covariates on the right-hand side of \code{regression_formula}.}

\item{response_weight_method}{A character string specifying the error 
distribution and link function to be used in the generalized additive model 
which estimates subjects' response propensity. By default, 
\code{response_weight_method} is set to \code{binomial(link = logit)}, which 
fits a logistic regression model. Any family function that \code{gam} accepts 
can be specified here.}

\item{treatment_weight_formula}{An object of class \code{formula} (or one that 
can be coerced to that class); a symbolic description of the model to be 
fitted. \code{treatment_weight_formula} specifies the model for estimating 
respondents' treatment propensity, which is used in weighting treatment effect
estimates that account for non-random attrition. By default, the formula 
includes only the confounder covariates included on the right-hand side of 
\code{regression_formula}. When accounting for attrition on unobservables,
response propensity weights are also included on the right-hand side of the 
formula.}

\item{treatment_weight_method}{A character string specifying the error 
distribution and link function to be used in the generalized additive model 
which estimates subjects' response propensity. By default, 
\code{response_weight_method} is set to \code{binomial(link = logit)}, which 
fits a logistic regression model. Any family function that \code{gam} accepts 
can be specified here.}

\item{n_bootstraps}{Numeric value defining the number of non-parametric 
bootstrap samples to be computed. The default number of bootstrap 
replications is 1000.}

\item{quantiles}{Vector of two numeric values between 0 and 1 specifying the 
quantiles at which non-parametric bootstrapped confidence intervals are to be 
returned. By default, \code{quantiles} are set to 0.05 and 0.95, which 
corresponds with a 90\% confidence interval.}

\item{n_cores}{Numeric value indicating the number of cores to allocate to 
running the function.  See \code{parallel} for details.}
}
\value{
A list of five elements containing the following:
 \item{coefficients}{A vector containing the estimated treatment effect, 
 bootstrapped standard error, and a bootstrapped confidence interval.}
 \item{weights}{A data.frame containing the response propensity weights, the
 treatment propensity weights, and the inverse probability weights used to
 estimate the treatment effect.}
 \item{effect}{A string noting the \code{effect_type} specified for the given
 estimate.}
 \item{attrition}{A string noting the \code{attrition_type} specified for the 
 given estimate.}
 \item{formulae}{A list containing the precise formulae that specify the 
 treatment effect function, the response propensity function, and the treatment 
 propensity function.}
}
\description{
\code{ipwlm} estimates the average treatment effect --- either among 
respondents (ATE|R = 1) or among the population (ATE) --- under various 
conditions of non-random attrition (attrition due to treatment alone, 
attrition due to treatment conditioned on observable confouding, or attrition 
due to treatment conditioned on observable and unobservable confounding).
}
\details{
The function estimates the treatment effect given non-random attrition, and 
uses non-parametric bootstrapping to estimate uncertainty.
}
\references{
Huber, Martin (2012): "Identification of Average Treatment Effects 
in Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler and David Miller. Special thanks to Jonas Markgraf and Hyunjoo Oh.
}
